Update Provinces_Database.sql
In this File, the author created a query, in which encompasses dataset for 34 Provinces in Indonesia. To begin with, this query comprises Province ID (Created by author), Province name, Island, alongside the corresponding capital for each provinces. Furthermore, the query encapsulated School Participation Rate (SPR), Human Development Index (HDIndex), and Cooperation Dividend in Million of Rupiah. To conclude, the data as for 2020 and it is sourced from National Bureau of Statistics (Badan Pusat Statistik or BPS Indonesia).